Testing normalization (#259)
This is a first pass at working out some testing kinks.

Previously we were using a combination of Node's assert, Jest's expect,
Jasmine's assert and Jasmine's expect.  We were using type definitions
appropriate for Node but not Jasmine.  This pass standardizes on expect.
It still uses Jasmine's expect in the browser; fixing that will come later.
This normalization is a necessary precursor step.

There were also issues with how we tested logging that required us to set
environment variable TZ=UTC.  This commit includes a framework for log testing
that streamlines tests and removes the need to override the system timezone.