Mocha/Chai instead of Jest/Jasmine for matter.js
Previously we were using Jest for node and Jasmine for the web.

This meant we could only use the subset of the Jest API that was identical to
the Jasmine API, and editor type support would lie as it only uses Jest types.
Instead use Mocha which is well supported for both Node and the Web.

This will also allow us to test the code we build rather than an intermediate
form generated by testing frameworks, although that'll be a subsequent commit.

Also added logic to hide log messages for successful tests.  This makes it much
easier to find failures and related log messages.

Does not convert submodules other than matter.js to Mocha.