Add ClusterStatus field for Status Responses (#336)
* Add ClusterStatus field for Status Responses

The StatusResponseError class got extended to also contain and transport a clusterStatus code for write and invoke responses. The handling in InteractionClient got enhanced to throw a StatusResponse error in such a case with both codes.
Testing will be later included via chip tool tests because here invoke cases are tested. For Write no cluster is implemented with a clusterStatusCode.

* add missing code parts

* add missing test adjustments