Add initializer and destroyer for ClusterServerHandlers (#337)
* Add initializer and destroyer for ClusterServerHandlers

Some clusters need more complex logic than just some functions and also require e.g. timer to be stopped when the ClusterServe is no longer needed or the device is stopped.
This PR introduces two new commandHandlers that can optionally be used. One to initialize the cluster Handlers - this one also gets attributes and events of the cluster. The second one is to destroy the logic and cleanup used resources.

* Add testing