Add DataEncoder to process DataReport data
This file introduces functions and also "intermediate types" that are near to DataReport types, but contain the schema and JS value instead of the already encoded value, so that chunking and compression can be applied dynamically.

InteractionServer and SubscriptionHandler (the places  that generate DataReports right now) will use these types instead of the final Tlv ones to prepare the data.
The provided functions here support dynamic atray chunking and tag compression