Add FailSafeManager class
This class instance will be by the "Device" once a Failsafe context is opened (armFailsafe) and will then be used to coordinate all Failsafe relevant topics. It maintains the Failsafe timer and handles expiry logic, it contains the "FabricBuilder" instance of the Fabric to build (or update), it will store relevant data to restore if the failsafe timer expires and such.
The commit also includes tests for general logic of the class. The Fabric and commission related tests are covered by chip tool tests (but could be replicated here later too, added todo)