Add update/revoke FabricIndex
Methods no longer needed are removed.
Add method to update a Fabric and also to "revoke" a fabric. The latter is used when during commissioning the fabric was already added but then the failsafe timer expired or other error cases occurred. Then the fabric is removed and if the fabric index was the last one generated we also reset the counter because the index was never officially used.