Make sure Event initial-data and updates are processed correctly
There was a variable name overlapping which caused the effect that events were not updated. Additionally, eventReports were not returned on subscription seed.
The test was hiding this behaviour because another event update came in from another global subscription and so was handles as ok but indeed was not. test adjusted to check behaviour. An additional test (testing event responses from initial subscription seed will be added in next PR because of code conflicts.