Move actual close session to ExchangeManager
Session closes can be triggered "synchronously" or "delayed when current running intercations were finished". The latter was not be correct because it is not the interaction that matters but the exchange that gets finished after the interaction is finished. Additionally the logic only implemented for MatterDevice nd not for Controller. So we move closing session logic to exchangemanager.