Introduce Node state
This commit is introducing a node state and a corresponding callback to inform the developer about state changes of the node. Additionally all callbacks are enhanced by the nodeId of the node the change happens. With this also generic callbacks can be provided when connecting to all nodes at once and the result can still be handled by node.
Node state also supports "structure changed" info when a bridge structure changed and was updated internally to reprocess the device objects if needed.