Promote tooling to first-class package
Moves "tools" into "packages/matter.js-tools".

Renames "run" and "build" to "matter-run" and "matter-build".

Adds tooling documentation.

Removes weird "file" dependency on self from tooling.  Not sure what the
purpose of it was but it appears to interfere with the installation of
scripts in node_modules/.bin when published on NPM.

Resolves #323