Handle Message extensions and Secured extensions correctly (#552)
* Handle Security/Message extensions correctly

Security and message extensions are read from the message data and provided but ignored as defined in specs right now. They can not be encoded.
Message decoding should also be more performant because it now do not encode the packet header again to get the securityFlags but use the original data.

* Optimize approach

* Tweak texts

* fix test