Add CAT support to Certificate managers
Add support for CATs in generation of NOCs including relevant validation. This also adds support for CATs on DER level and ASN.1 encoding - supporting multiple same subject tags is done here a bit more pragmatic because it is only about encoding, but added TODO for correct solution if needed again.