Refactor calculation of next session id
The logic to find the next available session id should wrap and search free spots - in worst case it should close the oldest unused session.
Additionally this commit moves the session id finding to the respective place where it is needed to prevent taking session ids that are then not ued because of errors.