Revert 0.7 APIs
Now that so little is shared between NodeServer and CommissioningServer it no
no longer makes sense to have a common base class.  Many of the changes now
were just moving things around, and there were breaking differences in the API
that are no longer necessary.

This reverts most of the changes to MatterNode, CommissioningServer,
CommissioningController and MatterServer.

CommissioningServer still has a number of internal changes as it integrates
components that do have changes.  But I don't think any of these changes
affect the public API.