Commissioning fixes (#656)
* Correctly respond on PAKE param errors

We monkey patch the Tlv definition to disable tlv schema validations for PAKE parameters when opening a commissioning window. Additionally we handle the relevant cases in code to return the right error as defined in specifications

* Expire Failsafe timer when revoking commissioning window

see https://github.com/project-chip/connectedhomeip/issues/31629

* Correctly respond Unsupported Access error

... when a fabric scoped command is called without associated fabric in commissioning cases

* Return error on commissioning complete with PASE session

* Code optimization

* Do not end commissioning mode when adding a fabric (addNoc)

According to specs commissioning mode ends with commissioning Complete but not before. Also to make sure that the PASE session is still intact.

* Catch socket close errors to prevent crash