Discovery and Session Refactoring to match 1.2 specs better (#636)
* Update SessionInterval Broadcastings

This commits adjusts the Session interval fields according to Matter 1.2, and adds SAT and ICD (always 0) to Broadcasts

* Restructure Session classes

This PR restructures session classes from an interface to an abstract base class. also the "Interval" parameters are renamed and enhanced to Matter 1.2.

* use new interval variable names

* Enhance MessageReceptionState class structure

... used in new session structure

* Adjust Tlv messages to match Matter 1.2 structure

* Refactor Operational device scanning

This commit contains a restructuring for operational and commissionable discovery. It now not just returns addresses but an Operational/Commissional device object including the addresses and discovery details. It also adds handling of TXT records for operational announcements.

* Enhance stored node details

* Adjust BleScanner interface to match with adjusted Scanner

* Use discover details in Case/Pase

This commit enhances the Case/Pase sessions to use discovery data for session intervals but also other sources - also store in resumption record.

* Enhance Node details also with basic information data after connection

* Use discovery data in controller

* Adjust to new methods

* Adjust tests

* Changelog

* add missing rename

* changelog

* Update CHANGELOG.md

* formatting

* merge

* changelog

* Address review comments

* further optimizations