Terminate on construction errors by default
If a behavior crashes the part now crashes.  Inconvenient override point by subclassing Part but not sure there's a
use case for this so not exposing at higher level.

If a part crashes the node now crashes.  May override by subclassing or adding observer to "partCrash" event and
returning false.