Devices-Gen test Fixes and Updates (#694)
* Cluster adjustments from tests

Some of them are device-gen fixes, some adoptions from master

* needed to prevent locking issues

I got an error when it was executed by the #commissioningTimeout worker

* Faabric scoped !== sensitive

* Correctly initialize isCommissioned flag

* Destroy ServerStore on factory Reset because will get reinitialized

* Add convenient "startUp" method for ServerNode

it is like "run" but resolves already when node is online. We can change the name but "start" was already used, so I decided for this for now. Lets talk

* Add info on react/callback

* Prevent danging promise ...

Or do we have a better way in the meantime?

* Removed unused settings

* Initialize Part Name

* Separate Init- and Defaultvalue in Attributeservers

... and handle initialization errors with fallback

* Adjust test to adjusted BasicInformation definition

In a former commit we adjusted the BasicInformation type to include the events ... with this these testcases do not work anymore. I did npt found any way to get them working again.

* Needed Adjustments to Legacy cluster impl

* Finalize/Fix OnOffServer implementation

* Implement Identify logic (no legacy)

* Add has method to VariableService

* Streamline mdns interface settings to one

* Document Environment variables

* Set log format to ANSI when tty and no config provided

... as legacy logic

* use getter for hasSignalSupport

* A Closed Json storage can be reinitialized

* Former commit needs Data model version exported

* Enhance logging

* Updated package-lock

* Add test scripts for devices-gen logic

* Fix

* Adjust Nocs/Fabrics to be fabric scoped

* Fixes

* Address review feedback