Fixes
First, behaviors now derive from:
    *.Cluster.with(<all non-exclusive features>)
instead of
    *.Complete

The complete cluster was kind of a hack and doesn't have the proper extension metadata to support with() correctly.  This would result in us incorrectly indicating support for elements.  It also includes conditional rules that are redundant and sometimes incorrect based on the limited metadata structure.

Second, in ValidatedElements detection of unsupported commands was a bit broken.  We'd warn a command wasn't supported if it was selected by feature but was optional.  And we'd then incorrectly register as supported.