Improve FabricIndex handling for Client usages (#674)
* Inject "empty" FabricIndex in Client set

... because if missing the validator might return an error

* Make Tlv encoding writeInteraction aware

In Write interactions the fabricIndex field must be omitted. But we can not omit the whole field because else our schema-aware validation and encoding returns errors. So we add a flag for encoding that do not encode the FabricIndex field.

* Adjust FabricIndex datatype with conditional validation

The FabricIndex datatype accepts when the "OMIT_FABRIC" magic constant is used.

* Add tests

* Changelog