Prevent async reaction to sync observables
Previously there was a "TODO" that warned it might be bad if this happened.  Turned out it was.  This adds a runtime
error to prevent this.  Could also probably prevent via types but will be a bit of a PITA so will wait to see if it's
a significant issue.