Fixes 312 (#769)
* Fixes 312

The PR fixes the PASE establishment issue, replaces elliptic and BN deps by @noble/curves

* Address review feedback

* make linter happy

* Implement node-style module resolution

Could have just used import.meta.resolve except esbuild throws a warning when transpiling to CJS and doesn't seem to
have a good way to disable.  We don't use it with CJS so would've been fine but not with warning.

So instead reimplement module resolution.  Use this to resolve the library list in web.ts.

---------

Co-authored-by: Greg Lauckhart <greg@lauckhart.com>