Spec ingestion refactor
Previously we scanned the TOC when ingesting clusters, then used that information to scan subsets of the document
associated with each cluster.  The newer implementation for ingesting devices OTOH just scans the entire document.
That latter turned out to be both less complex and more accurate.

This rewrites cluster ingestion to align with the device ingestion approach.

Includes a bit of cleanup and several fixes to the model, though only spec.ts is updated as yet.