Conditionally enable Reachable event on BasicInfo (#824)
* Conditionally enable Reachable event on BasicInfo

When the BasicInformation Cluster has the "reachable" attribute defined we automatically enable the corresponding event

* Enable reachable field in chip tests

* Remove test TestBasicInformation

... in new APIs, becaus ethats just an chiptool internal test... The cert tests do the tests in detail.

Reason: The TestBasicInformation has the hard expectation that reachable is not present as attribute and ignores PICS. The legacy API still executes this test

* Remove test TestBasicInformation

... also on legacy because same PICs file :-)

* Run cluster validation against behavior instance

This allows metadata in global attributes to reflect structural modifications made on a per-instance basis.  Immediate
use case is so BasicInformationServer can enable reachableChange event dynamically if reachable attribute is present.

* Finalize

* revert unwanted change

* revert unwanted change

---------

Co-authored-by: Greg Lauckhart <greg@lauckhart.com>