Implement AccessControl Server for Legacy
In the case of legacy the implementation just contains the logic of the cluster and is wired in the InteractionServer (see next commit) into the read/write/invoke calls.
Validation is handled in custom validators and event sending in custom setters. The getters and setters (because required by legacy api) need to handle itself the fabric scoped loading and saving. The new methods for this are also added in one of the next commits.