Persist also writable and fabric scoped data in new API
The legacy API was persisting also all writable and fabric scoped data. This commit does the same for the new API, else no ACL list value would ever be persisted :-)

In fact the matter specification is unclear here and tells that "any non-volatile data" must be persisted ... or other where applicable (kind of). So writable is obvious ...