Add a minimalistic GroupKeyCluster implementation
Some ACL related tests we need to pass also do stuff using group key cluster commands we do not had implemented.

So this adds a minimalistic Group key management implementation that just supports the IPK and not other groups.
But it already starts to wire some things into fabric where it also need to be later.